Make ContainsKeyRange return false when start is equal to end
ContainsKeyRange used to return true when a range has ["a", "b") and
start == end == "b". It seems this is not an expected behavior since
end key is exclusive.

This commit also fixes AdminMerge() and test code to pass (key,
key.Next()) to look up a range with a single key.

It might also make sense to check a given start/end key is valid in client.Scan, etc.